go back to using screenEvent
