+++ patches/net/minecraft/client/KeyboardHandler.java.patch
@@ -50,9 +50,9 @@
      public static void submitPreeditEvent(GuiEventListener element, @Nullable PreeditEvent event) {
          try {
 -            element.preeditUpdated(event);
-+            if (net.neoforged.neoforge.client.ClientHooks.onPreeditPre(element, event)) return;
++            if (element instanceof Screen screen && net.neoforged.neoforge.client.ClientHooks.onScreenPreeditPre(screen, event)) return;
 +            if (element.preeditUpdated(event)) return;
-+            net.neoforged.neoforge.client.ClientHooks.onPreeditPost(element, event);
++            if (element instanceof Screen screen) net.neoforged.neoforge.client.ClientHooks.onScreenPreeditPost(screen, event);
          } catch (Throwable t) {
              CrashReport report = CrashReport.forThrowable(t, "IME pre-edit event handler");
              if (element instanceof Screen screen) {
