+++ src/client/java/net/neoforged/neoforge/client/event/InputEvent.java
@@ -7,10 +7,8 @@ package net.neoforged.neoforge.client.event;
 
 import com.mojang.blaze3d.platform.InputConstants;
 import net.minecraft.client.KeyMapping;
-import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.input.KeyEvent;
 import net.minecraft.client.input.MouseButtonInfo;
-import net.minecraft.client.input.PreeditEvent;
 import net.minecraft.world.InteractionHand;
 import net.neoforged.bus.api.Event;
 import net.neoforged.bus.api.ICancellableEvent;
@@ -18,7 +16,6 @@ import net.neoforged.fml.LogicalSide;
 import net.neoforged.neoforge.common.NeoForge;
 import org.jetbrains.annotations.ApiStatus;
 import org.joml.Vector2ic;
-import org.jspecify.annotations.Nullable;
 import org.lwjgl.glfw.GLFW;
 
 /**
@@ -28,7 +25,6 @@ import org.lwjgl.glfw.GLFW;
  * @see InputEvent.MouseButton
  * @see MouseScrollingEvent
  * @see Key
- * @see Preedit
  * @see InteractionKeyMappingTriggered
  */
 public abstract class InputEvent extends Event {
@@ -302,62 +298,6 @@ public abstract class InputEvent extends Event {
          */
         public int getModifiers() {
             return this.keyEvent.modifiers();
-        }
-    }
-
-    /// Fired when an [Input Method Editor (IME) preedit event][PreeditEvent] is submitted to a GUI input target.
-    ///
-    /// An IME lets users enter text that is not directly available from their keyboard, such as Chinese, Japanese,
-    /// and Korean characters. While input is being composed, the IME supplies temporary preedit text that can change
-    /// until the user commits it.
-    ///
-    /// See [Pre][Preedit.Pre] and [Post][Preedit.Post] for listening before and after the normal handling.
-    public static abstract class Preedit extends InputEvent {
-        private final GuiEventListener inputTarget;
-        private final @Nullable PreeditEvent preeditEvent;
-
-        @ApiStatus.Internal
-        public Preedit(GuiEventListener inputTarget, @Nullable PreeditEvent preeditEvent) {
-            this.inputTarget = inputTarget;
-            this.preeditEvent = preeditEvent;
-        }
-
-        /// {@return the GUI input target that will receive the preedit event}
-        public GuiEventListener getInputTarget() {
-            return inputTarget;
-        }
-
-        /// {@return the current preedit composition, or `null` if there is no active composition}
-        public @Nullable PreeditEvent getPreeditEvent() {
-            return preeditEvent;
-        }
-
-        /// Fired **before** the preedit event is handled by the input target.
-        ///
-        /// This event is [cancellable][ICancellableEvent]. If the event is cancelled, the input target's preedit
-        /// handler will be bypassed and the corresponding [Post][Preedit.Post] will not be fired.
-        ///
-        /// This event is fired on the [main NeoForge event bus][NeoForge#EVENT_BUS],
-        /// only on the [logical client][LogicalSide#CLIENT].
-        public static class Pre extends Preedit implements ICancellableEvent {
-            @ApiStatus.Internal
-            public Pre(GuiEventListener inputTarget, @Nullable PreeditEvent preeditEvent) {
-                super(inputTarget, preeditEvent);
-            }
-        }
-
-        /// Fired **after** the preedit event is handled, if not handled by the input target
-        /// and the corresponding [Pre][Preedit.Pre] is not cancelled.
-        ///
-        /// This event is not [cancellable][ICancellableEvent].
-        ///
-        /// This event is fired on the [main NeoForge event bus][NeoForge#EVENT_BUS],
-        /// only on the [logical client][LogicalSide#CLIENT].
-        public static class Post extends Preedit {
-            @ApiStatus.Internal
-            public Post(GuiEventListener inputTarget, @Nullable PreeditEvent preeditEvent) {
-                super(inputTarget, preeditEvent);
-            }
         }
     }
